+++ src/tko/guide.py
@@ -1,0 +1,25 @@
+simple_guide = """
+╭─────────────────────── BASH ────────────────────────╮
+│Mostrar arquivos  : ls                               │
+│Mudar de pasta    : cd _nome_da_pasta                │
+│Subir um nível    : cd ..                            │
+│Apagar um arquivo : rm _nome_do_arquivo              │
+│Apagar uma pasta  : rm -r _nome_da_pasta             │
+│Renomear          : mv _antigo_nome _novo_nome       │
+│Mover             : mv _origem _destino              │
+╰─────────────────────────────────────────────────────╯
+╭─────────────────────── TKO ─────────────────────────╮
+│Atualizar replit  : update.sh                        │
+├─────────────────────────────────────────────────────┤
+│Baixar o problema : tko down [fup|poo|ed] _id _ext   │
+│ Exemplo para poo : tko down poo 012 cpp             │
+├─────────────────────────────────────────────────────┤
+│Rodar teste: tko run _codigo[s] _testes [-v] [-i ind]│
+│    exemplo: tko run solver.cpp cases.tio            │
+│  diff vert: tko run solver.cpp cases.tio -v         │
+│ só teste 6: tko run solver.cpp cases.tio -i 6       │
+├─────────────────────────────────────────────────────┤
+│Executar sem testar: tko exec _codigo                │
+│        exemplo    : tko exec solver.ts              │
+╰─────────────────────────────────────────────────────╯
+"""