+++ changelog.md
@@ -2,6 +2,9 @@
 
 - 0.0.9
   - removed tab from wdir.resume
+  - fixed type in whitespace
+  - added mismatch arrow
+  - added -g option to show simple guide
 - 0.0.8
   - add colors in diff inside the line
 - 0.0.7